switch replies api with comments api
