+++ src/Interfaces/http/api/comments/routes.js
@@ -1,8 +1,16 @@
 const routes = (handler) => ([
   {
     method: 'POST',
-    path: '/threads/{threadId}/comments/{commentId}/replies',
-    handler: handler.postReplyHandler,
+    path: '/threads/{threadId}/comments',
+    handler: handler.postCommentHandler,
+    options: {
+      auth: 'forumapi_jwt',
+    },
+  },
+  {
+    method: 'DELETE',
+    path: '/threads/{threadId}/comments/{commentId}',
+    handler: handler.deleteCommentHandler,
     options: {
       auth: 'forumapi_jwt',
     },
