+++ src/Interfaces/http/api/replies/routes.js
@@ -1,16 +1,8 @@
 const routes = (handler) => ([
   {
     method: 'POST',
-    path: '/threads/{threadId}/comments',
-    handler: handler.postCommentHandler,
-    options: {
-      auth: 'forumapi_jwt',
-    },
-  },
-  {
-    method: 'DELETE',
-    path: '/threads/{threadId}/comments/{commentId}',
-    handler: handler.deleteCommentHandler,
+    path: '/threads/{threadId}/comments/{commentId}/replies',
+    handler: handler.postReplyHandler,
     options: {
       auth: 'forumapi_jwt',
     },
